+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next (YYYY-MM-DD)
 
+- Use `XtermClear` on iTerm on macOS when `TERM` starts with `xterm`, to work around macOS not
+  having the correct terminfo by default.
+
 ## v1.0.0 (2021-04-25)
 
 Initial release